software_engineering/back-end/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
captain-pycard==0.9.11
cffi==1.14.0
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
numpy==1.18.1
packaging==20.1
pandas==1.0.1
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.0.0
